% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{add_diag}
\alias{add_diag}
\title{add_diag}
\usage{
add_diag(gp, dv)
}
\arguments{
\item{gp}{An H2-matrix.}

\item{dv}{A vector to be added to the diagonal.}
}
\value{
Takes gp, and adds dv to the diagonal in place.
}
\description{
Add provided vector 'dv' to diagonal of provided H2 matrix gp.
}
\examples{

}
