% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{cholesky_hmatrix}
\alias{cholesky_hmatrix}
\title{cholesky_hmatrix}
\usage{
cholesky_hmatrix(gp)
}
\arguments{
\item{gp}{An H-matrix, likely created by create_h2_kernel or equivalent.}
}
\value{
Cholesky decomposition of input.
}
\description{
Convert a matrix into its Cholesky form.
}
\examples{
 A <- matrix(data = rnorm(1000),
             nrow = 500,
             ncol = 2)
 gp_h2 <- create_h2_kernel(X = as.matrix(A),
                           hp = as.matrix(c(16, 1), ncol = 1),
                           kt = 1)
 gp_h  = convert_h2_to_h_matrix(gp = gp_h2)
 cholesky_hmatrix(gp = gp_h)

}
