% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{convert_h2_to_h_matrix}
\alias{convert_h2_to_h_matrix}
\title{convert_h2_to_h_matrix}
\usage{
convert_h2_to_h_matrix(gp)
}
\arguments{
\item{gp}{An H2-matrix.}
}
\value{
The converted form of the matrix.
}
\description{
Given an H2 matrix 'gp', convert it to a H matrix.
}
\examples{

}
