% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{create_h2_kernel}
\alias{create_h2_kernel}
\title{create_h2_kernel}
\usage{
create_h2_kernel(X, hp, kt)
}
\arguments{
\item{X}{Matrix of X coordinates to use to form the GP kernel.}

\item{hp}{Matrix of parameters for the Gaussian Process relative to the
current chosen kernel.}

\item{kt}{Kernel Type: the type of kernel to use as a number from the list of
current supported kernels: gaussian, gaussian_truncate, matern and diagonal.}
}
\value{
An H2 matrix with the specified kernel.
}
\description{
Return a pointer to a new instance of a h2_lib object.
}
\examples{
 # Note the use
 A <- matrix(data = rnorm(1000),
             nrow = 500,
             ncol = 2)
 gp_h2 <- create_h2_kernel(X = as.matrix(A),
                           hp = as.matrix(c(16, 1), ncol = 1),
                           kt = 1)
}
