% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{get_n_levels}
\alias{get_n_levels}
\title{get_n_levels}
\usage{
get_n_levels(N, M)
}
\arguments{
\item{N}{Integer.}

\item{M}{Integer.}
}
\value{
Returns log(N / M) / log(2).
}
\description{
Given integers N and M, computes n levels.
}
