% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{getSymmetricFactor}
\alias{getSymmetricFactor}
\title{getSymmetricFactor}
\usage{
getSymmetricFactor(GPobj)
}
\arguments{
\item{GPobj}{The H matrix, assembled and factorized for use.}
}
\value{
Extracts the symmetric factor from the Gaussian Process
compressed object.
}
\description{
Extract symmetric factor from an H matrix
}
\details{
Build and return the symmetric factor W from an H matrix, already
assembled and factorized.
}
