% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HODLR_sample.R
\name{HODLR_TP_sample}
\alias{HODLR_TP_sample}
\title{HODLR_TP_sample}
\usage{
HODLR_TP_sample(Y, TP, x, cur_tau, l_scale, idx = NULL, kernel = "matern")
}
\arguments{
\item{Y}{The observed data.}

\item{TP}{This parameter represents g. In practice, this value can represent
multiple dimensions.}

\item{x}{Values of the covariate associated with f.}

\item{cur_tau}{Current precision estimate of \eqn{Y}, i.e., \eqn{Y = f\times g + \epsilon}.
where \eqn{\epsilon \sim N(0,\tau^-1)}.}

\item{l_scale}{The length scale parameter}

\item{idx}{This optional parameter is the index of the x sorted in increasing order. 
If it is not specified, the function will sort x for the Gibbs sampler. 
However, since most samplers only require x to be sorted once, 
this is not efficient.   We recommend passing the index (i.e. order(x)).}

\item{kernel}{The kernel function to use. By default the "matern" kernel is used but the 
squared exponential kernel "sqr-exp" is also available.}
}
\description{
Assuming a tensor product \eqn{f*g}, this function samples f given g 
using a Gibbs sampler defined in Moran and Wheeler.
}
\examples{

}
