% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HODLR_sample.R
\name{HODLR_TP_sample_ls}
\alias{HODLR_TP_sample_ls}
\title{HODLR_TP_sample_ls}
\usage{
HODLR_TP_sample_ls(
  ls,
  Y,
  TP,
  x,
  cur_tau,
  bounds,
  idx = NULL,
  kernel = "matern"
)
}
\arguments{
\item{ls}{The current value of the length scale being used for f.}

\item{Y}{The observed data or the residual given other components in
the mean.}

\item{TP}{This parameter represents g. In practice, this value can represent
multiple dimensions.}

\item{x}{Values of the covariate associated with f.}

\item{cur_tau}{Current precision estimate of \eqn{Y}, i.e., \eqn{Y = f\times g + \epsilon}.
where \eqn{\epsilon \sim N(0,\tau^-1)}.}

\item{bounds}{Uniform distribution [a,b] specified for prior distribution on length scale.}

\item{idx}{This optional parameter is the index of the x sorted in increasing order. 
If it is not specified, the function will sort x for the Gibbs sampler. 
However, since most samplers only require x to be sorted once, 
this is not efficient.   We recommend passing the index (i.e. order(x)).}

\item{kernel}{The kernel function to use. By default the "matern" kernel is used but the 
squared exponential kernel "sqr-exp" is also available.}
}
\description{
Assuming a tensor product \eqn{f*g}, this function samples the length-scale
for f, given the marginal distribution of Y given g. It performs
a simple Metropolis step where the proposal distribution is 
a uniform(-0.5,0.5) and the prior distribution on the length-scale
is a uniform distribution specified by the bounds argument.
}
\examples{

}
