% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_det_HODLR}
\alias{log_det_HODLR}
\title{log_det_HODLR}
\usage{
log_det_HODLR(GPobj)
}
\arguments{
\item{GPobj}{The H matrix, assembled and factorized for use.}
}
\value{
\eqn{\log |A|} where \eqn{|A|} is the determinant of the matrix.,
}
\description{
Returns the log-determinant of a  HODLR matrix
}
\details{
Givne a HODLR matrix A returns \eqn{\log | A |}
}
