% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{log_sqrt_determinant_hmatrix}
\alias{log_sqrt_determinant_hmatrix}
\title{log_sqrt_determinant_hmatrix}
\usage{
log_sqrt_determinant_hmatrix(gp)
}
\arguments{
\item{gp}{An H-matrix, likely created by create_h2_kernel() or equivalent.}
}
\value{
Log determinant of the input H-matrix.
}
\description{
Compute the log determinant of an H-matrix
}
\details{
Compute the log-determinant of the input H-matrix. Note: no square root
is taken at any point in this process, so name is confusing.
}
