% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{matmatProduct}
\alias{matmatProduct}
\title{matmatProduct}
\usage{
matmatProduct(GPobj, M)
}
\arguments{
\item{GPobj}{The H matrix, assembled and factorized for use.}

\item{M}{A matrix of doubles, as dynamically accessed via Eigen.}
}
\value{
Computes the matrix-matrix product of GPobj and M, which works
even if M is a N*1 matrix.
}
\description{
Compute a matrix-matrix product
}
\details{
Compute a matrix-matrix or matrix-vector product of the abstracted
kernel matrix and the provided matrix M. Overloads the matmatProduct()
function from the Eigen package.
}
