% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiply_h2_stdnorm}
\alias{multiply_h2_stdnorm}
\title{multiply_h2_stdnorm}
\usage{
multiply_h2_stdnorm(gp)
}
\arguments{
\item{gp}{An H2-matrix.}
}
\value{
The multiplication gp * rv. Note that the seed in the global environment
  is used, as the random normal generation internally to this function
  is just rnorm().
}
\description{
Multiply an H2 matrix by Standard Normals
}
\details{
Given an H2 matrix 'gp', right-multiply it by a vector 'rv' of N(0,1) random numbers.
}
\examples{

}
