% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiply_h2_vector}
\alias{multiply_h2_vector}
\title{multiply_h2_vector}
\usage{
multiply_h2_vector(gp, dv)
}
\arguments{
\item{gp}{An H2-matrix.}

\item{dv}{A vector to by multiplied by gp, that is, gp * dv.}
}
\value{
Takes gp and dv and computes the multiplication gp * dv.
}
\description{
Left-multiply provided vector 'dv' by provided H2 matrix gp.
}
\examples{

}
