% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{multiply_hmat_vec}
\alias{multiply_hmat_vec}
\title{multiply_hmat_vec}
\usage{
multiply_hmat_vec(h, v)
}
\arguments{
\item{h}{An H-matrix.}

\item{v}{A vector.}
}
\value{
The matrix product of h*v
}
\description{
Multiply an H-matrix by a vector.
}
