% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{read_h_matrix}
\alias{read_h_matrix}
\title{read_h_matrix}
\usage{
read_h_matrix(file)
}
\arguments{
\item{file}{System file reference or file name.}
}
\value{
rv An H-matrix.
}
\description{
Read an H-matrix from disk in NetCDF binary format.
}
\details{
Read an H-matrix from disk in NetCDF binary format. Partner function
to save_h_matrix().
}
\examples{



}
