% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h2_gp_single_sampler.R
\name{sample_gp}
\alias{sample_gp}
\title{sample_gp}
\usage{
sample_gp(h2, h, y, tX, sigma_p, tau)
}
\arguments{
\item{h2}{H^2 approximation matrix to the covariance matrix.}

\item{h}{Cholesky decomposition of the covariance matrix}

\item{y}{For solving}

\item{tX}{Regression coefficients or weights}

\item{sigma_p}{Precision of the variance-covariance matrix}

\item{tau}{Data precision.}
}
\description{
sample_gp
}
