% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{save_h_matrix}
\alias{save_h_matrix}
\title{save_h_matrix}
\usage{
save_h_matrix(file, gp)
}
\arguments{
\item{file}{System file reference or file name.}

\item{gp}{An H-matrix, likely created by create_h2_kernel or equivalent.}
}
\value{
Nothing.
}
\description{
Save an H-matrix to disk in NetCDF binary format.
}
