% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup_compressedMatrixGP.R
\name{setup_compressedMatrixGP}
\alias{setup_compressedMatrixGP}
\title{Given a specified kernel (currently supporting matern and squared Exponential),
and a matrix of locations X, set up a compressed matrix Gaussian process
pointer via h2lib.}
\usage{
setup_compressedMatrixGP(X, kernel_type = "matern", kernel_params, tol, M)
}
\arguments{
\item{X}{Matrix of locations to use.}

\item{kernel_type}{String specifying which kernel to use. Defaults to matern; currently supported are 'matern' and 'squared_exponential'.}

\item{kernel_params}{Named list of parameters for the chosen kernel. For supported kernels at present, must contain both 'sigma' and 'rho'.}

\item{tol}{Specified tolerance for accuracy of calculations.}

\item{M}{Maximum submatrix size in the H^2 decomposition.}
}
\value{
ptr Pointer to the compressed approximation matrix Gaussian Process.
}
\description{
Given a specified kernel (currently supporting matern and squared Exponential),
and a matrix of locations X, set up a compressed matrix Gaussian process
pointer via h2lib.
}
