% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setup_compressedMatrixGP_Matern}
\alias{setup_compressedMatrixGP_Matern}
\title{setup_compressedMatrixGP_Matern}
\usage{
setup_compressedMatrixGP_Matern(X, sigma, rho, tol, M)
}
\arguments{
\item{X}{Matrix of locations.}

\item{sigma}{Function standard deviation.}

\item{rho}{Length-Scale.}

\item{tol}{Tolerance for accuracy of calculations.}

\item{M}{Maximum sub-matrix size.}
}
\value{
After setting up compressed memory storage, returns a pointer
 to a HODLR_Tree object. Specifically, in this case, the Matern kernel
 (default) is the case of $p=2$, or $nu = 5/2$.

   K(r) = sigma^2 * [1 + sqrt(5) * r / rho + 5/3 * (r / rho)^2] * exp[-sqrt(5) * r / rho]

 For this implementation, the diagonal elements for the nugget are set to 1e-8.
}
\description{
Set Up Compressed Gaussian Process with Matern Kernel
}
\details{
Given a matrix of locations X, standard deviation sigma, and length
scale rho, sets up a compressed matrix for a Gaussian process using
a Matern kernel, with maximum submatrix size M. Default is the case $p=2$.
}
