% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setup_compressedMatrixGP_Matern_tP}
\alias{setup_compressedMatrixGP_Matern_tP}
\title{setup_compressedMatrixGP_Matern_tP}
\usage{
setup_compressedMatrixGP_Matern_tP(X, tP, sigma, rho, tol, M)
}
\arguments{
\item{X}{Matrix of locations.}

\item{tP}{Specified kernel.}

\item{sigma}{Function standard deviation.}

\item{rho}{Length-Scale.}

\item{tol}{Tolerance for accuracy of calculations.}

\item{M}{Maximum sub-matrix size.}
}
\value{
After setting up compressed memory storage, returns a pointer
 to a HODLR_Tree object. Specifically, in this case, the Matern_pX_Kernel
 refers to a kernel such that $p=2$, i.e.,

   K(r) = sigma^2 * (1 + sqrt(5) * r / rho + 5/3 * (r / rho)^2) * exp(-sqrt(5) * r / rho)

 but with the nugget diagonals being set to 1/P, for P a N*1 matrix.
}
\description{
Set Up Compressed Gaussian Process with Matern (tP) Kernel
}
\details{
Given a matrix of locations X, standard deviation sigma, and length
scale rho, sets up a compressed matrix for a Gaussian process using
a Matern kernel, with maximum submatrix size M. This is $p=2$, with
user-provided reciprocal diagonal components for the nugget.
}
