% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setup_compressedMatrixGP_MaternP1}
\alias{setup_compressedMatrixGP_MaternP1}
\title{setup_compressedMatrixGP_MaternP1}
\usage{
setup_compressedMatrixGP_MaternP1(X, sigma, rho, tol, M)
}
\arguments{
\item{X}{Matrix of locations.}

\item{sigma}{Function standard deviation.}

\item{rho}{Length-Scale.}

\item{tol}{Tolerance for accuracy of calculations.}

\item{M}{Maximum sub-matrix size.}
}
\value{
After setting up compressed memory storage, returns a pointer
 to a HODLR_Tree object. Specifically, in this case, the Matern_p1_Kernel
 refers to a kernel with $p=1$ and $nu=3/2$:

   K(r) = σ^2 * [1 + sqrt(3) * r / ρ] * exp[-sqrt(3) * r / ρ]

 For this implementation, the diagonal elements for the nugget are set to 1.
}
\description{
Set Up Compressed Gaussian Process with Matern (P1) Kernel
}
\details{
Given a matrix of locations X, standard deviation sigma, and length
scale rho, sets up a compressed matrix for a Gaussian process using
a Matern kernel, with maximum submatrix size M. This is the $p=1$ case.
}
