% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{setup_compressedMatrixGP_sqrExp_tP}
\alias{setup_compressedMatrixGP_sqrExp_tP}
\title{setup_compressedMatrixGP_sqrExpP1}
\usage{
setup_compressedMatrixGP_sqrExp_tP(X, tP, sigma, rho, tol, M)
}
\arguments{
\item{X}{Matrix of locations.}

\item{sigma}{Function standard deviation.}

\item{rho}{Length-Scale.}

\item{tol}{Tolerance for accuracy of calculations.}

\item{M}{Maximum sub-matrix size.}
}
\value{
After setting up compressed memory storage, returns a pointer
 to a HODLR_Tree object. Specifically, in this case, kernel is the
 Squared Exponential Kernel:

   K(r) = sigma^2 * exp[ -rho * ||r||^2 ]

 with the nugget diagonals set to 1.0, as with the Matern p1 case.
}
\description{
Set Up Compressed Gaussian Process with Square Exponential (P1) Kernel
}
\details{
Given a matrix of locations X, standard deviation sigma, and length
scale rho, sets up a compressed matrix for a Gaussian process using
the Squared Exponential kernel.
}
