% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{simulate_compressedMatrixGP}
\alias{simulate_compressedMatrixGP}
\title{simulate_compressedMatrixGP}
\usage{
simulate_compressedMatrixGP(N, GPobj)
}
\arguments{
\item{N}{Number of observations to draw.}

\item{GPobj}{The H matrix, assembled and factorized for use.}
}
\value{
Given N and GPobj, generate N random N(0, 1)s, then
   use HODLR code to do a quick symmetric factor product. Returns
   the product, which are the draws from the process using compression.
}
\description{
Simulate a draw from a compressed GP.
}
