% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_h2_vector_start}
\alias{solve_h2_vector_start}
\title{solve_h2_vector_start}
\usage{
solve_h2_vector_start(gp, d_precond, b, start, tolerance, iter)
}
\arguments{
\item{gp}{An H2-matrix}

\item{d_precond}{A vector of pre-conditions.}

\item{b}{A vector b, for equation $gp*x = b$, with dimension matching that of gp.}
}
\value{
Given inputs gp and b, solves for x in $Ax=b$, and returns it.
}
\description{
Solve Ax=b using H2 matrices with initial conditions.
}
\details{
Solve a linear system $Ax=b$ with the preconditioned generalized minimal
residual method from H2lib, with provided initial condition.
Note that the underlying H2 class function defaults to
iter = 2000 and eps = 1e-14 (eps is overwritten by argument tolerance).
}
\examples{



}
