% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_hmatrix}
\alias{solve_hmatrix}
\title{solve_hmatrix}
\usage{
solve_hmatrix(gp, x)
}
\arguments{
\item{gp}{An H-matrix, likely created by create_h2_kernel() or equivalent.}

\item{b}{A vector of constants.}
}
\value{
Solution of matrix operation, x.
}
\description{
Solve Ax=b, for A an H-matrix
}
\details{
Solve a matrix equation Ax=b for x, where A is approximated in H-matrix form.
}
\examples{
 A <- matrix(data = rnorm(1000),
             nrow = 500,
             ncol = 2)
 gp_h2 <- create_h2_kernel(X = as.matrix(A),
                           hp = as.matrix(c(16, 1), ncol = 1),
                           kt = 1)
 gp_h  = convert_h2_to_h_matrix(gp_h2)
 cholesky_hmatrix(gp_h)
 b <- matrix(rnorm(500), nrow = 500, ncol = 1)
 x <- solve_hmatrix(gp_h, b)

}
