% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{solve_HODLR}
\alias{solve_HODLR}
\title{solve_HODLR}
\usage{
solve_HODLR(GPobj, b)
}
\arguments{
\item{GPobj}{The H matrix, assembled and factorized for use.}

\item{b}{A matrix (vector), the right-hand side of the Ax=b equation.}
}
\value{
Solves the matrix equation Ax=b for an A matrix which is already
assembled and factorized by HODLR.
}
\description{
Solve Ax=b using HODLR
}
\details{
Solves the matrix equation Ax=b by using the kernel object's abstraction
inverse, i.e., A^{-1} * b.
}
