% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HODLR_sample.R
\name{HODLR_TP_sample_scale}
\alias{HODLR_TP_sample_scale}
\title{HODLR_TP_sample_cale}
\usage{
HODLR_TP_sample_scale(Y, TP, cur_tau, mean = 0, prec = 1)
}
\arguments{
\item{Y}{The observed \eqn{n \times 1} data vector or the residual given \
other components in the mean.}

\item{TP}{A matrix of dimension \eqn{n \times k}. Here, each row corresponds
to an observation in Y and each column corresponds to one of the k dimensions 
in the tensor Product.}

\item{cur_tau}{Current precision estimate of \eqn{Y}, i.e., \eqn{Y = \gamma f_1 \cdot f_2\cdots f_k + \epsilon}
where \eqn{\epsilon \sim N(0,\tau^-1)}.}

\item{mean}{Mean for scale, default 0.}

\item{prec}{Precision argument, default 1.}
}
\description{
Assuming a tensor product of k 1-D functions, i.e. \eqn{\gamma f_1 \cdot f_2\cdots f_k},
 this function samples \eqn{\gamma}, which is the scale parameter. 
 Here \eqn{\gamma \sim N(\mu,\alpha^-1)}, and by default \eqn{\mu = 0} and \eqn{\alpha =1}.
}
